api/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
pydantic>=2.8.0
pydantic-settings>=2.4.0
sqlalchemy>=2.0.30
httpx>=0.27.0
psycopg[binary]>=3.2.0
redis>=5.0.0
PyJWT>=2.9.0
passlib>=1.7.4
pytest>=8.2.0
